Fix for input objects not within bounds of imgae.
git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@6738 2422e80f-9519-0410-a819-8e21e7c86b6d

+++ src/gary/Laguerre.cpp
@@ -36,7 +36,7 @@ LVector::LVector(const DVector& rhs, int order_):
 }
 
 LVector::LVector(const CVector& cv) {
-  int ord = floor(0.5*(sqrt(8.*cv.size()+1.)-3.));
+  int ord = int(floor(0.5*(sqrt(8.*cv.size()+1.)-3.)));
   if (PQIndex::size(ord) != int(cv.size())) 
     throw LaguerreError("CVector size incompatible with LVector");
   // ??? should also check for Hermiticity
